Remove "normalization" of search input appearance
This rule comes from the necolas/normalize.css reset, with the comment:
"Correct the odd appearance in Chrome and Safari.".

An older version of the comment said:
"Address `appearance` set to `searchfield` in Safari 5 and Chrome."

The README says:
"The search input is not fully stylable by default.
In Chrome and Safari on OSX/iOS you can't control font, padding, border, or background.
In Chrome and Safari on Windows you can't control border properly.
It will apply border-width but will only show a border color (which cannot be controlled) for the outer 1px of that border.
Applying -webkit-appearance: textfield addresses these issues without removing the benefits of search inputs (e.g. showing past searches)."
Link: https://github.com/necolas/normalize.css/blob/master/README.md#typesearch

I've tested it in all three major browsers (Chrome, Firefox, and Safari) and it is now fully customizable, so I'm sure that this rule is outdated and thus no longer necessary.
Demo: https://codepen.io/germanfrelo/pen/PwYpLBW

Additionally, the unprefixed 'appearance' property is already supported in all major browser engines (and the default value is 'auto'):
- caniuse.com: https://caniuse.com/?search=appearance
- Blink (Chromium/Chrome) has it since version 84 (2020):
  - https://issues.chromium.org/issues/41459076
  - https://issues.chromium.org/issues/41460175
  - https://239d877-dot-cr-status.appspot.com/feature/4715298156445696
- WebKit (Safari) has it since version 15.4 (2022):
  - https://bugs.webkit.org/show_bug.cgi?id=143842
  - https://bugs.webkit.org/show_bug.cgi?id=231534
  - https://bugs.webkit.org/show_bug.cgi?id=236012
- Gecko (Firefox) has it since version 80 (2020):
  - https://bugzilla.mozilla.org/show_bug.cgi?id=1620467

Related to the input search element:
Chrome and Safari have a clear/cancel button in input search elements, but Firefox has it as an experimental feature that can be enabled, but is not enabled by default:
- https://developer.mozilla.org/en-US/docs/Mozilla/Firefox/Experimental_features#layout_for_input_typesearch
- https://bugzilla.mozilla.org/show_bug.cgi?id=1654288
- https://bugzilla.mozilla.org/show_bug.cgi?id=1417753
- https://bugzilla.mozilla.org/show_bug.cgi?id=1481608
- https://github.com/w3c/csswg-drafts/issues/5365