Remove font-weight normalization of b & strong
The Web Specification recommends rendering 'b' and 'strong' elements with 'font-weight: bolder':
https://html.spec.whatwg.org/multipage/rendering.html#phrasing-content-3

Before 2025-02-20, Blink/Chromium and WebKit/Safari were using 'bold',
and Gecko/Firefox was the only major browsers using 'bolder'.

After that, Blink/Chromium changed to use 'bolder':
https://github.com/germanfrelo/base-css-stylesheet/commit/361d07c347afa746ba180101fa7940e517906e98

And as of 2025-03-09, WebKit has changed its behavior and now uses 'bolder' as well,
to align with the rest of the browser engines:
https://github.com/WebKit/WebKit/commit/2f7d04cbae67d35d6c3f2830cb134313c9b0dde1

So this normalization is no longer needed.